// src/prospects/dto/create-prospect.dto.ts
export class CreateProspectDto {
  personneId?: string;
  nom?: string;
  prenom?: string;
  telephone?: string;
  quartier?: string;
  dateNaissance?: string;
  lieuNaissance?: string;
  tuteurNom?: string;
  tuteurPrenom?: string;
  tuteurTelephone?: string;
  filiereSouhaitee?: string;
  intention?: string;
  statutRelance?: string;
}